backend/requirements.txt
requests==2.26.0
django==2.2.6
djangorestframework==3.11.0
drf-extra-fields==3.2.1
djoser==2.1.0
pillow==8.2.0
python-dotenv==0.13.0
django-filter==21.1
asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1